requirements-dev.txt
# Development dependencies for AWS KMS Platform
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
bandit>=1.7.0
safety>=2.3.0

# Pre-commit hooks
pre-commit>=3.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=1.0.0

# Development tools
ipython>=8.0.0
jupyter>=1.0.0
notebook>=6.5.0

# Debugging
pdbpp>=0.10.0
ipdb>=0.13.0

# Performance profiling
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Security scanning
bandit>=1.7.0
safety>=2.3.0
semgrep>=1.0.0

# API testing
httpx>=0.24.0
pytest-httpx>=0.24.0

# Database testing
pytest-sqlalchemy>=0.1.0

# Mock AWS services
moto>=4.2.0
boto3-stubs>=1.26.0

# Code coverage
coverage>=7.0.0
coveralls>=3.3.0

# Development server
watchdog>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Type checking
types-requests>=2.28.0
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0

# Additional utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.64.0 